Abstract new utils
