+++ scripts/release/publish.mjs
@@ -1,9 +1,4 @@
-import fs from "node:fs";
-import path from "node:path";
-import {exitIf, help, shell} from "./utils.mjs";
-
-// TODO: Allow to customize
-const workspacesPath = "./packages";
+import {exitIf, help, resolveMonorepoDirpaths, shell} from "./utils.mjs";
 
 const tag = process.argv[2];
 
@@ -17,10 +12,6 @@ See https://github.com/ChainSafe/lodestar/blob/unstable/RELEASE.md#1-create-rele
 `);
 exitIf(!tag, "<tag> not set");
 
-for (const dirname of fs.readdirSync(workspacesPath)) {
-  const dirpath = path.join(workspacesPath, dirname);
-  const pkgPath = path.join(dirpath, "package.json");
-  if (fs.existsSync(pkgPath)) {
-    shell(`npm publish --tag ${tag}`, {cwd: dirpath});
-  }
+for (const dirpath of resolveMonorepoDirpaths()) {
+  shell(`npm publish --tag ${tag}`, {cwd: dirpath});
 }
