+++ scripts/release/set_version.mjs
@@ -1,14 +1,11 @@
 import fs from "node:fs";
 import path from "node:path";
-import {exitIf, help} from "./utils.mjs";
+import {cleanVersion, exitIf, help, resolveMonorepoDirpaths} from "./utils.mjs";
 
 /* eslint-disable no-console, @typescript-eslint/no-unsafe-return, @typescript-eslint/explicit-function-return-type */
 
-// TODO: Allow to customize
-const workspacesPath = "./packages";
-
 // Must clean version, could have 'v' prefix
-const version = (process.argv[2] ?? "").replace(/^v/, "");
+const version = cleanVersion(process.argv[2] ?? "");
 const range = process.argv[3];
 const rangeValues = ["caret", "tilde", "exact"];
 
@@ -55,19 +52,19 @@ forEachPkg((pkg) => {
  * @param {(pkg: PkgJson) => PkgJson | undefined} cb
  */
 function forEachPkg(cb) {
-  for (const dirname of fs.readdirSync(workspacesPath)) {
+  const dirpaths = resolveMonorepoDirpaths();
+
+  for (const dirpath of dirpaths) {
     try {
-      const pkgPath = path.join(workspacesPath, dirname, "package.json");
-      if (fs.existsSync(pkgPath)) {
-        const pkg = readPkg(pkgPath);
-        const newPkg = cb(pkg);
-        if (newPkg) {
-          // Append new line to reduce diff when modifying manually
-          fs.writeFileSync(pkgPath, JSON.stringify(newPkg, null, 2) + "\n");
-        }
+      const pkgPath = path.join(dirpath, "package.json");
+      const pkg = readPkg(pkgPath);
+      const newPkg = cb(pkg);
+      if (newPkg) {
+        // Append new line to reduce diff when modifying manually
+        fs.writeFileSync(pkgPath, JSON.stringify(newPkg, null, 2) + "\n");
       }
     } catch (e) {
-      e.message = `Error on package ${dirname}: ${e.message}`;
+      e.message = `Error on package ${dirpath}: ${e.message}`;
       throw e;
     }
   }
